fix: refactor accounts management menu item
